+++ jest.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+    "verbose": true,
+    "testMatch": [__dirname  + '/gdbgui/src/js/tests/**']
+}